requirements.txt
python-dotenv
black
tabulate